ep4-python-solc/requirements.txt
py-solc-x==2.0.3
